Rename 'awapi_accessor' to 'autoware'
