+++ external/awapi_accessor/CMakeLists.txt
@@ -13,7 +13,7 @@ find_package(ament_cmake_auto REQUIRED)
 
 ament_auto_find_build_dependencies()
 
-ament_auto_add_library(${PROJECT_NAME} SHARED src/awapi_accessor.cpp)
+ament_auto_add_library(${PROJECT_NAME} SHARED src/autoware.cpp)
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
